Fuzzed file fix.

+++ plugins/excel/ms-obj.c
@@ -1302,8 +1302,8 @@ ms_read_OBJ (BiffQuery *q, MSContainer *c, MSObjAttrBag *attrs)
 	}
 
 	obj->excel_type_name = NULL;
-	if (obj->excel_type < (int)G_N_ELEMENTS (object_type_names))
-		obj->excel_type_name = object_type_names [obj->excel_type];
+	if ((size_t)obj->excel_type < G_N_ELEMENTS (object_type_names))
+		obj->excel_type_name = object_type_names[obj->excel_type];
 	if (obj->excel_type_name == NULL)
 		obj->excel_type_name = "Unknown";
